requirements.txt
blis==0.7.9
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
confection==0.0.4
cymem==2.0.7
decorator==4.4.2
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.4.1/en_core_web_lg-3.4.1-py3-none-any.whl
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.1/en_core_web_sm-3.4.1-py3-none-any.whl
en-core-web-trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.4.1/en_core_web_trf-3.4.1-py3-none-any.whl
es-core-news-lg @ https://github.com/explosion/spacy-models/releases/download/es_core_news_lg-3.4.0/es_core_news_lg-3.4.0-py3-none-any.whl
es-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/es_core_news_sm-3.4.0/es_core_news_sm-3.4.0-py3-none-any.whl
filelock==3.9.0
huggingface-hub==0.12.0
idna==3.4
imageio==2.24.0
imageio-ffmpeg==0.4.8
Jinja2==3.1.2
joblib==1.2.0
langcodes==3.3.0
MarkupSafe==2.1.1
moviepy==1.0.3
murmurhash==1.0.9
nltk==3.8.1
numpy==1.24.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==23.0
pandas==1.5.2
pathy==0.10.1
Pillow==9.4.0
preshed==3.0.8
proglog==0.1.10
pydantic==1.10.4
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
six==1.16.0
smart-open==6.3.0
spacy==3.4.4
spacy-alignments==0.9.0
spacy-legacy==3.0.11
spacy-loggers==1.0.4
spacy-transformers==1.1.9
SpeechRecognition==3.9.0
srsly==2.4.5
thinc==8.1.7
timecode==1.3.1
tokenizers==0.13.2
torch==1.13.1
tqdm==4.64.1
transformers==4.25.1
typer==0.7.0
typing_extensions==4.4.0
urllib3==1.26.14
wasabi==0.10.1

opencv-python~=4.7.0.68